Move mock GraphQL object functions to @livepeer/graphql-sdk to be imported by other packages
