+++ packages/explorer/src/views/Account/enhance.js
@@ -2,13 +2,16 @@ import React from 'react'
 import { compose, withHandlers } from 'recompose'
 import { graphql } from 'react-apollo'
 import gql from 'graphql-tag'
-import { MathBN, mockAccount } from '../../utils'
+import { MathBN } from '../../utils'
+import { utils as gqlSDKUtils } from '@livepeer/graphql-sdk'
 import {
   connectCoinbaseQuery,
   connectCurrentRoundQuery,
   connectToasts,
   withTransactionHandlers,
 } from '../../enhancers'
+
+const { mockAccount } = gqlSDKUtils
 
 const accountQuery = gql`
   fragment AccountFragment on Account {
