+++ packages/explorer/src/views/Transcoders/enhance.js
@@ -3,7 +3,10 @@ import { compose, mapProps, withHandlers } from 'recompose'
 import { graphql } from 'react-apollo'
 import gql from 'graphql-tag'
 import { connectCurrentRoundQuery, connectToasts } from '../../enhancers'
-import { MathBN, mockAccount, sleep } from '../../utils'
+import { MathBN, sleep } from '../../utils'
+import { utils as gqlSDKUtils } from '@livepeer/graphql-sdk'
+
+const { mockAccount } = gqlSDKUtils
 
 const MeDelegatorTranscoderQuery = gql`
   fragment DelegatorFragment on Delegator {
